log: Fix build warning
Fix the following build warning on arm32:

/reference/src/log.c: In function 'log_message':
/reference/src/log.c:99:41: warning: format '%ld' expects argument of type 'long int', but argument 4 has type '__time64_t' {aka 'long long int'} [-Wformat=]
   99 |         written = snprintf(p, len, "[%8ld.%9ld]: [%s]: ", time.tv_sec, time.tv_nsec,
      |                                      ~~~^                 ~~~~~~~~~~~
      |                                         |                     |
      |                                         long int              __time64_t {aka long long int}
      |                                      %8lld

+++ src/log.c
@@ -96,7 +96,7 @@ void log_message(enum log_level level, const char *format, ...)
 	len = sizeof(buffer) - 1;
 	p = (char *)buffer;
 
-	written = snprintf(p, len, "[%8ld.%9ld]: [%s]: ", time.tv_sec, time.tv_nsec,
+	written = snprintf(p, len, "[%8lld.%9ld]: [%s]: ", (long long int)time.tv_sec, time.tv_nsec,
 			   log_level_to_string(level));
 	p += written;
 	len -= written;
